mcp-servers/fileserver/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.2
fastmcp==2.0
pyyaml==6.0.1
python-dotenv==1.0.0
uvicorn[standard]>=0.35.0
websockets>=15.0.1
requests>=2.31.0
cryptography>=3.4.8
